feat: update env and config
